Don't run conda-docker on Mac
conda-docker doesn't work on Mac, and making it work is not straightforward,
so for now we just disable it when we aren't on Linux.

With this and #537, basic Mac support seems to work with --standalone (#507).
I'm not sure what else I should be testing, but I can create environments and
they appear to be functional.