Make tests use a temporary directory for conda environments
Previously they used the current conda prefix, but this is dangerous (there's
a risk a broken test could break the dev environment), and the current conda
prefix could contain things that break the tests.

This currently works by parameterizing in the fixture itself (currently just
for two basic envs, one without pip packages and one with). But in the future
we can do this parameterization on a per-test level with indirect
parameterization. See
https://docs.pytest.org/en/latest/example/parametrize.html#indirect-parametrization

Currently, the test_generate_conda_export() test fails here. The reason is
that it creates an export which cannot actually be read by the
CondaSpecification object.

The reason for this is that the "name" of the environment that gets generated
is just the same as the "prefix". This fails a validation check, because the
name is guarded by a regular expression that disallows paths.

However, this appears to be a bug in conda-store, or maybe even an
architectural issue. The "name" of a conda environment, as defined by "conda
env export" at least, is only really defined if that environment is in the
envs directory. I checked, and even env export yamls generated by conda-store
itself (in standalone mode) end up having the prefix for the name, rather than
the supposed actual environment name.

At best, it means that this test is just wrong and it shouldn't actually check
the generated YAML file by trying to load it into a CondaSpecification.

I don't know if conda actually stores the environment "name" anywhere in the
environment. It seems to me that this data should be stored separately by
conda-store, especially since the environment path for environments created by
conda-store is not actually the same as the environment name (the path
contains a hash).