Fix test_generate_conda_export locally
The environment name produced by conda env export will not be the actual
"name" of the environment unless that environment is inside of a conda envs
directory. Otherwise it is just the prefix, which fails the CondaSpecification
validation for the name. We work around this by just hotfixing the name in the
test before parsing the result.