Fixup: support both aliases in `role_mappings`
One of the previous commits explicitly defined both aliases in
`role_mappings`, but that was wrong because only one value is used in
the DB, which is `developer`. Instead, you need to map `developer` from
the DB to `role_mappings`, which can either have `editor` (new name) or
`developer` (legacy name). The new name is checked first, then the old
one.