Remove exception handler that prints to the console
Errors will already be printed to the console if the debug level
is set to debug.

Having this exception handler catch all exceptions and invoke
the http_exception_handler is incorrect behaviour. This causes
the http_exception_handler to be called for non HTTPExceptions.
Since the exception passed in is the wrong type, this often
causes a new exception to be triggered.